Created a model to store course activity data storing time spend time by user on courses
