+++ lms/djangoapps/course_home_api/models.py
@@ -23,3 +23,13 @@ class DisableProgressPageStackedConfig(StackedConfigurationModel):
         return "DisableProgressPageStackedConfig(disabled={!r})".format(
             self.disabled
         )
+
+class CourseActivityLog(models.Model):
+    id = models.AutoField(primary_key=True)
+    user_id = models.IntegerField(max_length=225)
+    course_id = models.CharField(max_length=300)
+    start_time = models.DateTimeField(auto_now_add=True)
+    end_time = models.DateTimeField(default=None, null=True)
+
+    def __str__(self):
+        return self.user_id
